refactor : #19 폴더 구조 변경
- 고객용과 사장님용 폴더를 분리해서 개발하기 위해 생성
