+++ lib/owner/screens/letter_screen.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:carousel_slider/carousel_slider.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:get/get.dart';
 
 class LetterPage extends StatefulWidget {
   const LetterPage({super.key});
