+++ lib/owner/screens/receipt_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:frontend/screens/review_screen.dart';
-import 'package:frontend/widgets/store_widget.dart';
+import 'package:frontend/owner/screens/review_screen.dart';
+import 'package:frontend/owner/widgets/store_widget.dart';
 import 'package:get/get.dart';
 
 class ReceiptPage extends StatefulWidget {
